Extract Page(De)Serializer interface
